accepts and tabulates votes
I am currently working w/ my own system - the important part is that it initiates a handshake client side the second it connects to allow it to proceed, and then the votes are tallied later. A lot of this has been copied and pasted directly from my AAT_stag_hare project (under ethans repo) so redirect any questions that way.