added generations
realized the probability matrix was only getting calcualted once, which was a problem, so I have no changed it to NOT do that so every bot calculates it on its own. this makes it take around 10 times longer than it should (slighly less as we are able to avoid some overhead). I'm gonna look for more optimizations but I might just need to rewrite it in cpp.