made a new algorithm
i mean thats all there is to it - implemented the "better greedy" algorithm, which is just greedy but with expected value. it has a better? cooperation score than greedy or whatnot, so thats a start. And it only takes 5.25 seconds to run 10,000 games. see the docs for all the juicy deets.